hypervisor: rename source file to be more general
Some new features are needed within the hypervisor system schema. Rename
source file to reflect this.
